Reorder controls on line/spectrum toolbars
