+++ cypress/specs/app.spec.ts
@@ -1,4 +1,4 @@
-const SNAPSHOT_DELAY = 100;
+const SNAPSHOT_DELAY = 500;
 
 describe('App', () => {
   beforeEach(() => {
