+++ src/h5web/vis-packs/models.ts
@@ -1,7 +1,6 @@
 import type { ElementType } from 'react';
 import type { IconType } from 'react-icons';
 import type { Entity } from '../providers/models';
-import type { ToolbarProps } from '../toolbar/Toolbar';
 
 export interface VisContainerProps {
   entity: Entity;
@@ -10,6 +9,6 @@ export interface VisContainerProps {
 export interface VisDef {
   name: string;
   Icon: IconType;
-  Toolbar?: ElementType<ToolbarProps>;
+  Toolbar?: ElementType;
   Container: ElementType<VisContainerProps>;
 }
